requirements.txt
Django==1.11.2
mysqlclient==1.3.10
PyMySQL==0.7.11
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
